refactor: remplacement de l'importation de PostType pour utiliser post.schema au lieu de post-type.enum
