--- apps/api/src/post/post-type.enum.ts
@@ -1,7 +1,0 @@
-export enum PostType {
-    Presentiel = 'Présentiel',
-    Teletravail = 'Télétravail',
-    Hybride = 'Hybride',
-}
-
-export default PostType;
