+++ apps/api/test/unit/post/post.service.spec.ts
@@ -4,7 +4,7 @@ import { Model, Types } from 'mongoose';
 import { PostService } from '../../../src/post/post.service';
 import { Post } from '../../../src/post/post.schema';
 import { CreatePostDto } from '../../../src/post/dto/createPost.dto';
-import { PostType } from '../../../src/post/post-type.enum';
+import { PostType } from '../../../src/post/post.schema';
 
 describe('PostService', () => {
     let service: PostService;
